backend/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Django==5.2
django-cors-headers==4.7.0
django-rest-knox==5.0.2
djangorestframework==3.16.0
Flask==3.1.0
flask-cors==5.0.1
frozendict==2.4.6
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multitasking==0.0.11
numpy==2.2.5
pandas==2.2.3
pandas_ta==0.3.14b0
peewee==3.17.9
platformdirs==4.3.7
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
setuptools==80.3.1
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
yfinance==0.2.59